- update to belongsTo(), hasOne() and hasMany() for the new argument joinKey. This argument allows the developer to specify a column to join to on the opposite side of the relationship from the foreignKey.
A great use case for this happened to me today where I was working with a legacy database and needed to model a join that was WAY outside of wheels conventions. The database could not be changed, so the feature was born.

I have added a couple of tests to prove out the code. It should be obvious that joins like these are not the intended purpose, but they prove the code works as expected.