transactions fixes
the following problems were identified with transactions:

the original transaction mode would not be respected during
during callbacks. as an example, if you set the model to
rollback and callback performed a create (which default to
commit), the created record would not get rolled back.

"none" transaction modes would never close and as such
subsequent transactions would not have mode respected.
as an example, if you performed a "none" transaction and
then performed a "rollback" transaction, the rollback
wouldn't be respected.