Setting to only include safe characters in raw SQL. This effectively makes it so the only thing that gets through is either stuff sent with cfqueryparam or safe SQL.
This could break backwards compatibility (if someone has a query that uses an operator that is not supported by Wheels and the SQL gets passed along with a character not in the safe list) so if we don't want this functionality in the next release we can set the setting to a blank string for now.

Something else to consider is that we could skip the strippping if parameterize=false too.