Add 'Railo4' servername environment variable
Railo 4.X has some breaking changes from 3.x. One of the most
notable changes is the way it handles the casing structure keys.
In 3.x the key were uppercased and in 4.x they are loweredcased.
This caused a bunch of tests to fail. By assigning the specific
version to the servername variable, we are able to get the tests
to pass.