fixes #126 casing with property labels
We were lower casing the property label and then capitalizing the entire
error message of the property based upon whether the property marker was
at the beginning of the message. This caused numerous problems when labels
contained two words like `Last Name` as it would be displayed as `Last
name`.

The solution is to really not do anything with the property labels in the
messages.