Made changes to fix the breaking test cases
The test cases that were breaking was due to the group by statement that does not add the calculated property if there is an aggregate function in the query to the group by clause. This was breaking the test case so added the condition to only throw error in case of "select". Also added another condition due to the breaking of self join test case. If the table is self joined, the same columns are already in the select clause so the "local.toAppend" does not get the value in case of duplicate columns from the same table.
It gets the columns if the dupes are in different tables as they are then aliased with table dot notation.