Issue 580 select ambiguous column name using the wheels alias (#1253)
* issue 580 select ambiguous column name using the wheels alias

Added a new aliasedProperyList node in the classData which will have the specific columns prepended with the model name so that we can get the column in the select clause if include table is specified.

* Removed condition for specific columns

Changed condition to include all columns in the list as suggested by @chapmandu

* Update select.cfc

Due to the changes done in this branch, the other test cases need to be altered a little. Fixed that.