Fixing Adobe 2018, 2021 and 2023
Modified the mappings in the application.cfc and onapplicationstart.cfc as it was working fine in lucee but was giving error in Adobe Coldfusion.
Added a third argument for "mid()" function as it is required in Adobe Coldfusion.
Adobe Coldfusion returns an object of "java.time.LocalDateTime" for datetime data. Modified the code to convert that into a datetime object only for Adobe Coldfusion.