Fix for test cases
Modified the code to convert java.time.LocalDateTime into a datetime object when it is returned. Also removed the Adobe coldfusion condtion. So now, even if lucee updates and does the same as adobe coldfusion and starts returning a java.time.LocalDateTime object then we won't have to update our code to accommodate for it.
Also added a condition specifically for adobe2018 for returning error.