Fixing lucee5 error and adding a dropdown for JSON format for testbox
Fixed the error for lucee5 that was causing the container to not start. Changed the Path from "/app/public" to "/app" in DockerFile.
Added a dropdown for the Testbox button. Now there are two format options for the testbox(HTML and JSON).
Added a condition in application.cfc that stops the HTML and CSS from debug.cfm to be returned with the JSON in testbox.