Adding Database param for test environment and header to JSON
Added a Database param in the code. Now if a db parameter is passed in the url that is one of the following: mysql, sqlserver, posgres and h2, then that database's datasource will be used.
Also added the header in JSON return format so the github actions can determine if the tests have passed or not.