Adding testbox for application level testing
Added the functionality to write testbox tests for your application
Added the "_assets" for so user can keep their assets for their tests there.
add a route in the "routes.cfm" for these tests
Added a routes.cfm file in the testbox folder so user can write their own routes for their application testing there.
Added a dummy test file "Example.cfc" to see the working of this functionality
Added an second dropdown in the "Testbox" button in the navbar. Now, the user can select between application tests or core test in all the available formats(HTML, JSON, TXT, JUnit)